% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ages.R
\name{add_ages}
\alias{add_ages}
\title{Add age columns}
\usage{
add_ages(split_data)
}
\arguments{
\item{split_data}{A named list of data frames, one per REDCap tab.}
}
\value{
`split_data` with the age columns added.
}
\description{
Adds `age_at_best_fev1` to `baseline_characteristics` and
`age_at_spirometry` to `visits`, in years since `partial_dob`.
}
