% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_baseline_char}
\alias{get_baseline_char}
\title{Extract baseline characteristics data}
\usage{
get_baseline_char(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of baseline characteristics data (everything not included in follow up visits)
}
\description{
Extract baseline characteristics data
}
