% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_demographics}
\alias{get_demographics}
\title{Extract demographics data}
\usage{
get_demographics(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of demographics data.
}
\description{
Extract demographics data
}
