% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_medications}
\alias{get_medications}
\title{Extract medication data}
\usage{
get_medications(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of medication data.
}
\description{
Extract medication data
}
