% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_pex}
\alias{get_pex}
\title{Extract pulmonary exacerbation data}
\usage{
get_pex(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of pulmonary exacerbation data.
}
\description{
Extract pulmonary exacerbation data
}
