% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_pregnancy}
\alias{get_pregnancy}
\title{Extract pregnancy data}
\usage{
get_pregnancy(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of pregnancy data.
}
\description{
Extract pregnancy data
}
