% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_side_effects}
\alias{get_side_effects}
\title{Extract side effect data}
\usage{
get_side_effects(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of side effect data.
}
\description{
Extract side effect data
}
