% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_visits}
\alias{get_visits}
\title{Extract visit data}
\usage{
get_visits(export)
}
\arguments{
\item{export}{A data frame containing the raw REDCap export.}
}
\value{
A data frame of visit data.
}
\description{
Extract visit data
}
\seealso{
[stack_visits()]
}
