% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_logger.R
\name{qc_logger}
\alias{qc_logger}
\title{Create a QC logger}
\usage{
qc_logger()
}
\value{
A list with `add()` to record a QC entry and `save_log()` to
  save the log to file.
}
\description{
Create a QC logger
}
