% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_api.R
\name{from_api}
\alias{from_api}
\title{USe the REDCap API to load a raw export}
\usage{
from_api(token)
}
\arguments{
\item{token}{API token to use}
}
\value{
A data frame of the raw export data
}
\description{
USe the REDCap API to load a raw export
}
